Update foxfooding.md (#720)
so anyhow - I needed to test out profiles and it wasn't in Firefox Labs to use.  

Then I remembered Nimbus Dev Tools for foxfooding and it was the SIMPLEST thing ever.  Santiago had feedback that my directions with videos were too long.  He was right.

I moved videos and verbose stuff to the end and just put the check list that takes 60 seconds to get set up to foxfood.